backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.9
pydantic>=1.10.7
psycopg2-binary>=2.9.6
alembic>=1.10.3
python-dotenv>=1.0.0
httpx>=0.24.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
spotipy>=2.22.1
python-socketio>=5.8.0
bcrypt>=4.0.1
